optimize files
